+++ core/utils/format.ts
@@ -8,3 +8,23 @@ const byteUnit = Intl.NumberFormat("en", {
 export function bytes(bytes: number) {
   return byteUnit.format(bytes);
 }
+
+const percentageUnit = Intl.NumberFormat("en", {
+  style: "percent",
+  minimumFractionDigits: 1,
+  maximumFractionDigits: 1,
+});
+
+export function percentage(initialValue: number, finalValue: number) {
+  if (initialValue === finalValue) {
+    return "0%";
+  }
+
+  const ratio = -(1 - finalValue / initialValue);
+
+  if (ratio < 0) {
+    return percentageUnit.format(ratio);
+  }
+
+  return "+" + percentageUnit.format(ratio);
+}