+++ CHANGELOG.md
@@ -9,13 +9,12 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [3.0.1] - Unreleased
 ### Added
 - Restore `site.copy()` function removed in 3.0.0.
-- Minify html plugin: output info in the debugbar
-- esbuild plugin: output input info in the debugbar
+- Info from minify_html, esbuild, terser, postcss and svgo plugins to the debugbar.
 
 ### Fixed
 - Warning message by the transform_images plugin.
 - Don't ignore `/.well-known` folders.
-- Updated dependencies: `tailwindcss`, `unocss` and some icons.
+- Updated dependencies: `tailwindcss`, `unocss`, `lume-bar` and some icons.
 
 ## [3.0.0] - 2025-05-07
 ### Added